refactor: added react query and tryCatch
Refactored the providers into their own file, while adding in React Query client provider. Also,
created a new tryCatch util function to add better typing to my try catch blocks in the future. This
is all in preparation for further commits on this branch where I'll be refactoring the entire app to
use react-query.

BREAKING CHANGE: All APIs are being re-written, and all components refactored, to utilise
react-query and tryCatch. This is massive, so I'm deeming it breaking.